Split config admin pages into components
