+++ frontend/src/pages/admin/AdminUsersPage.vue
@@ -1,0 +1,10 @@
+<template>
+  <MainLayout>
+    <UsersManagementCard />
+  </MainLayout>
+</template>
+
+<script setup lang="ts">
+import MainLayout from "@/layouts/MainLayout.vue";
+import UsersManagementCard from "@/components/admin/UsersManagementCard.vue";
+</script>
